+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Het gemiddelde cijfer van alle vakken wordt nu weergeven
+
 ### Changed
 - De footer neemt de hele breedte in en heeft een margin aan de onderkant